Use external session only via constructor in repositories
