+++ tests/unit_of_work/sync/test_session_lifecycle.py
@@ -25,19 +25,17 @@ def test_session_is_removed_on_cleanup(sa_manager):
 def test_commit_is_called_only_if_not_read_only(
     mocked_uow_commit: MagicMock,
     read_only_flag,
-    repository_class,
     model_class,
     sa_manager,
 ):
     uow = SessionHandler(sa_manager.get_bind())
-    repo1 = repository_class(sa_manager.get_bind())
 
     # Populate a database entry to be used for tests
     model1 = model_class(
         name="Someone",
     )
     with uow.get_session(read_only=read_only_flag) as _session:
-        repo1.save(model1, session=_session)
+        _session.add(model1)
 
     assert mocked_uow_commit.call_count == int(not read_only_flag)
 
@@ -51,7 +49,7 @@ def test_rollback_is_called_if_commit_fails(
 
     failure_exception = Exception("Some Error")
     mocked_session = MagicMock(spec=scoped_session)
-    uow._session = mocked_session
+    uow.session = mocked_session
     if commit_fails:
         mocked_session.commit.side_effect = failure_exception
